Scanning utilities and scanning Functions
